2 byte kind

+++ socks6.h
@@ -136,7 +136,7 @@ enum SOCKS6OperationReplyCode
 
 struct SOCKS6Option
 {
-	uint8_t  kind;
+	uint16_t kind;
 	uint16_t len;
 	uint8_t  data[0];
 } __attribute__((packed));
